fix: Mouse scrolling in VM (#291)
* parse amount with radix

This can lead to bugs if there’s a leading `0` or potentially other values.

* Don't override amount when using the mouse

Not sure why this was done, but it was surprising behavior when I explicitly set a large amount value.

* Send direction to e2b

https://e2b.dev/docs/sdk-reference/desktop-js-sdk/v1.9.2/sandbox#scroll supports up/down.

Sending -1 actually throws because a value has to be >= 1.

* Warn if scrolling didn't appear to work

This is pretty reliable in my testing, *unless* the time changes or something between events.

If anything, it helps when there are lots of scroll attempts and fails at least once.

* Log error for using unsupported types

* Wait 1 second after +f before typing